+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+- StanzaSlash add some stanza filter unit tests to fill in the checks
+
 ## 4.5.1 - *2022-06-05*
 
 Standardise files with files in sous-chefs/repo-management